backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scipy==1.14.1
six==1.16.0
tqdm==4.66.6
tzdata==2024.2
Werkzeug==3.0.4
